significant code update for setup
significant update to create setup stage
slight modification of css
added gfxhelper class to create arrowheads
modified displacement to use doubles instead of floats
other updates as part of the setup integration

+++ src/game/helpers/Displacement.java
@@ -4,9 +4,9 @@ import game.ui.PlanetManagerUI;
 
 public class Displacement
 {
-	private float x, y;
+	private double x, y;
 
-	public Displacement(float x, float y)
+	public Displacement(double x, double y)
 	{
 		this.x = x;
 		this.y = y;
@@ -22,31 +22,31 @@ public class Displacement
 		return new Displacement(this.x - d.x, this.y - d.y);
 	}
 
-	public float length()
+	public double length()
 	{
 		return (float) Math.sqrt(Math.pow(x, 2) + Math.pow(y, 2));
 	}
 
-	public Displacement add(float length)
+	public Displacement add(double length)
 	{
 		double angle = Math.atan2(y, x);
-		float x = (float) (this.x + length * Math.cos(angle));
-		float y = (float) (this.y + length * Math.sin(angle));
+		double x = this.x + length * Math.cos(angle);
+		double y = this.y + length * Math.sin(angle);
 		
 		return new Displacement(x, y);
 	}
 	
-	public Displacement subtract(float length)
+	public Displacement subtract(double length)
 	{
 		return add(-length);
 	}
 
-	public float getX()
+	public double getX()
 	{
 		return x;
 	}
 
-	public float getY()
+	public double getY()
 	{
 		return y;
 	}
